fix: split to standalone module
