+++ command-object/src/descriptor_pointer.rs
