+++ command-object/src/graphics_command.rs
@@ -1,309 +1,6 @@
-use std::ops::Range;
-
 use bedrock as br;
 
-#[repr(transparent)]
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct BufferUsage(pub u64);
-impl BufferUsage {
-    pub const UNUSED: Self = Self(0);
-
-    pub const HOST_RO: Self = Self(1 << 0);
-    pub const HOST_RW: Self = Self(1 << 1);
-    pub const TRANSFER_SRC: Self = Self(1 << 2);
-    pub const TRANSFER_DST: Self = Self(1 << 3);
-    pub const VERTEX_BUFFER: Self = Self(1 << 4);
-    pub const INDEX_BUFFER: Self = Self(1 << 5);
-    pub const INDIRECT_BUFFER: Self = Self(1 << 6);
-
-    // shader resources usage
-    pub const VERTEX_UNIFORM: Self = Self(1 << 8);
-    pub const VERTEX_STORAGE_RO: Self = Self(1 << 9);
-    pub const VERTEX_STORAGE_RW: Self = Self(1 << 10);
-    pub const FRAGMENT_UNIFORM: Self = Self(1 << 11);
-    pub const FRAGMENT_STORAGE_RO: Self = Self(1 << 12);
-    pub const FRAGMENT_STORAGE_RW: Self = Self(1 << 13);
-
-    pub const fn has_bits(&self, bits: Self) -> bool {
-        (self.0 & bits.0) != 0
-    }
-
-    pub fn vk_pipeline_stage_mask_requirements(&self) -> br::PipelineStageFlags {
-        let mut f = br::PipelineStageFlags(0);
-
-        f = if self.has_bits(Self::HOST_RO | Self::HOST_RW) {
-            f.host()
-        } else {
-            f
-        };
-        f = if self.has_bits(Self::TRANSFER_SRC | Self::TRANSFER_DST) {
-            f.transfer()
-        } else {
-            f
-        };
-        f = if self.has_bits(Self::VERTEX_BUFFER) {
-            f.vertex_input()
-        } else {
-            f
-        };
-        f = if self.has_bits(Self::INDEX_BUFFER) {
-            f.vertex_input()
-        } else {
-            f
-        };
-        f = if self.has_bits(Self::INDIRECT_BUFFER) {
-            f.draw_indirect()
-        } else {
-            f
-        };
-        f = if self
-            .has_bits(Self::VERTEX_UNIFORM | Self::VERTEX_STORAGE_RO | Self::VERTEX_STORAGE_RW)
-        {
-            f.vertex_shader()
-        } else {
-            f
-        };
-        f = if self.has_bits(
-            Self::FRAGMENT_UNIFORM | Self::FRAGMENT_STORAGE_RO | Self::FRAGMENT_STORAGE_RW,
-        ) {
-            f.fragment_shader()
-        } else {
-            f
-        };
-
-        f
-    }
-
-    pub fn vk_access_flags_requirements(&self) -> br::vk::VkAccessFlags {
-        let mut f = 0;
-
-        f |= if self.has_bits(Self::HOST_RO) {
-            br::AccessFlags::HOST.read
-        } else {
-            0
-        };
-        f |= if self.has_bits(Self::HOST_RW) {
-            br::AccessFlags::HOST.write | br::AccessFlags::HOST.read
-        } else {
-            0
-        };
-        f |= if self.has_bits(Self::TRANSFER_SRC) {
-            br::AccessFlags::TRANSFER.read
-        } else {
-            0
-        };
-        f |= if self.has_bits(Self::TRANSFER_DST) {
-            br::AccessFlags::TRANSFER.write
-        } else {
-            0
-        };
-        f |= if self.has_bits(Self::VERTEX_BUFFER) {
-            br::AccessFlags::VERTEX_ATTRIBUTE_READ
-        } else {
-            0
-        };
-        f |= if self.has_bits(Self::INDEX_BUFFER) {
-            br::AccessFlags::INDEX_READ
-        } else {
-            0
-        };
-        f |= if self.has_bits(Self::INDIRECT_BUFFER) {
-            br::AccessFlags::INDIRECT_COMMAND_READ
-        } else {
-            0
-        };
-        f |= if self.has_bits(Self::VERTEX_UNIFORM | Self::FRAGMENT_UNIFORM) {
-            br::AccessFlags::UNIFORM_READ
-        } else {
-            0
-        };
-        f |= if self.has_bits(Self::VERTEX_STORAGE_RO | Self::FRAGMENT_STORAGE_RO) {
-            br::AccessFlags::SHADER.read
-        } else {
-            0
-        };
-        f |= if self.has_bits(Self::VERTEX_STORAGE_RW | Self::FRAGMENT_STORAGE_RW) {
-            br::AccessFlags::SHADER.read | br::AccessFlags::SHADER.write
-        } else {
-            0
-        };
-
-        f
-    }
-}
-impl std::ops::BitOr for BufferUsage {
-    type Output = Self;
-
-    fn bitor(self, rhs: Self) -> Self::Output {
-        Self(self.0 | rhs.0)
-    }
-}
-impl std::ops::BitOrAssign for BufferUsage {
-    fn bitor_assign(&mut self, rhs: Self) {
-        self.0 |= rhs.0;
-    }
-}
-
-fn vk_pipeline_stage_mask_requirements_for_image_layout(
-    l: br::ImageLayout,
-) -> br::PipelineStageFlags {
-    match l {
-        br::ImageLayout::Undefined | br::ImageLayout::Preinitialized => br::PipelineStageFlags(0),
-        br::ImageLayout::General => br::PipelineStageFlags::ALL_COMMANDS,
-        br::ImageLayout::ColorAttachmentOpt => br::PipelineStageFlags::COLOR_ATTACHMENT_OUTPUT,
-        br::ImageLayout::DepthStencilAttachmentOpt => {
-            br::PipelineStageFlags::EARLY_FRAGMENT_TESTS.late_fragment_tests()
-        }
-        br::ImageLayout::DepthStencilReadOnlyOpt => br::PipelineStageFlags::EARLY_FRAGMENT_TESTS,
-        br::ImageLayout::ShaderReadOnlyOpt => {
-            br::PipelineStageFlags::VERTEX_SHADER.fragment_shader()
-        }
-        br::ImageLayout::TransferSrcOpt | br::ImageLayout::TransferDestOpt => {
-            br::PipelineStageFlags::TRANSFER
-        }
-        br::ImageLayout::PresentSrc => br::PipelineStageFlags::BOTTOM_OF_PIPE,
-    }
-}
-
-pub struct BufferUsageTransitionBarrier<'r, B: br::Buffer> {
-    pub buffer: &'r RangedBuffer<B>,
-    pub from_usage: BufferUsage,
-    pub to_usage: BufferUsage,
-}
-impl<B: br::Buffer> BufferUsageTransitionBarrier<'_, B> {
-    pub fn make_vk_barrier(&self) -> br::BufferMemoryBarrier {
-        br::BufferMemoryBarrier::new(
-            &self.buffer.0,
-            self.buffer.1.clone(),
-            self.from_usage.vk_access_flags_requirements(),
-            self.to_usage.vk_access_flags_requirements(),
-        )
-    }
-}
-impl<B: br::Buffer> PipelineBarrierEntry for BufferUsageTransitionBarrier<'_, B> {
-    fn add_into(self, barrier: &mut PipelineBarrier) {
-        barrier.buffer_barriers.push(self.make_vk_barrier());
-        barrier.src_stage_mask |= self.from_usage.vk_pipeline_stage_mask_requirements();
-        barrier.dst_stage_mask |= self.to_usage.vk_pipeline_stage_mask_requirements();
-    }
-
-    fn reserve_hints(barrier: &mut PipelineBarrier, count: usize) {
-        barrier.buffer_barriers.reserve(count);
-    }
-}
-impl<const N: usize, B: br::Buffer> GraphicsCommand for [BufferUsageTransitionBarrier<'_, B>; N] {
-    fn execute(self, cb: &mut br::CmdRecord<'_, impl br::CommandBuffer + br::VkHandleMut>) {
-        PipelineBarrier::new().with_barriers(self).execute(cb)
-    }
-}
-
-pub struct RangedBuffer<B: br::Buffer>(pub B, pub Range<u64>);
-impl<B: br::Buffer> RangedBuffer<B> {
-    pub const fn from_offset_length(buffer: B, offset: u64, length: usize) -> Self {
-        Self(buffer, offset..offset + length as u64)
-    }
-
-    pub const fn for_type<T>(buffer: B, offset: u64) -> Self {
-        Self::from_offset_length(buffer, offset, std::mem::size_of::<T>())
-    }
-
-    pub const fn offset(&self) -> u64 {
-        self.1.start
-    }
-
-    pub const fn byte_length(&self) -> u64 {
-        self.1.end - self.1.start
-    }
-
-    pub fn make_ref<'s>(&'s self) -> RangedBuffer<&'s B> {
-        RangedBuffer(&self.0, self.1.clone())
-    }
-
-    pub fn descriptor_uniform_buffer_write_info(&self) -> (br::vk::VkBuffer, Range<usize>) {
-        (
-            self.0.native_ptr(),
-            self.1.start as usize..self.1.end as usize,
-        )
-    }
-
-    pub fn usage_barrier(
-        &self,
-        from_usage: BufferUsage,
-        to_usage: BufferUsage,
-    ) -> BufferUsageTransitionBarrier<B> {
-        BufferUsageTransitionBarrier {
-            buffer: &self,
-            from_usage,
-            to_usage,
-        }
-    }
-
-    pub fn usage_barrier3(
-        &self,
-        first_usage: BufferUsage,
-        intermedial_usage: BufferUsage,
-        last_usage: BufferUsage,
-    ) -> [BufferUsageTransitionBarrier<B>; 2] {
-        [
-            self.usage_barrier(first_usage, intermedial_usage),
-            self.usage_barrier(intermedial_usage, last_usage),
-        ]
-    }
-
-    pub fn usage_barrier3_switching(
-        &self,
-        first_usage: BufferUsage,
-        intermedial_usage: BufferUsage,
-    ) -> [BufferUsageTransitionBarrier<B>; 2] {
-        self.usage_barrier3(first_usage, intermedial_usage, first_usage)
-    }
-
-    pub fn inner_ref(&self) -> &B {
-        &self.0
-    }
-}
-impl<B: br::Buffer + Clone> RangedBuffer<&'_ B> {
-    pub fn clone_inner_ref(&self) -> RangedBuffer<B> {
-        RangedBuffer(self.0.clone(), self.1.clone())
-    }
-}
-impl<B: br::Buffer + br::MemoryBound + br::VkHandleMut, M: br::DeviceMemory + br::VkHandleMut>
-    RangedBuffer<peridot::Buffer<B, M>>
-{
-    pub fn guard_map<R>(
-        &mut self,
-        action: impl FnOnce(&br::MappedMemoryRange<M>) -> R,
-    ) -> br::Result<R> {
-        self.0.guard_map(self.1.clone(), action)
-    }
-}
-
-pub struct RangedImage<R: br::Image>(R, br::ImageSubresourceRange);
-impl<R: br::Image> RangedImage<R> {
-    pub fn single_color_plane(resource: R) -> Self {
-        Self(resource, br::ImageSubresourceRange::color(0..1, 0..1))
-    }
-
-    pub fn barrier(
-        &self,
-        from_layout: br::ImageLayout,
-        to_layout: br::ImageLayout,
-    ) -> br::ImageMemoryBarrier {
-        br::ImageMemoryBarrier::new(&self.0, self.1.clone(), from_layout, to_layout)
-    }
-
-    pub fn barrier3(
-        &self,
-        first_layout: br::ImageLayout,
-        intermedial_layout: br::ImageLayout,
-        last_layout: br::ImageLayout,
-    ) -> [br::ImageMemoryBarrier; 2] {
-        [
-            self.barrier(first_layout, intermedial_layout),
-            self.barrier(intermedial_layout, last_layout),
-        ]
-    }
-}
+use crate::{vk_pipeline_stage_mask_requirements_for_image_layout, Mesh};
 
 pub trait GraphicsCommand {
     fn execute(self, cb: &mut br::CmdRecord<'_, impl br::CommandBuffer + br::VkHandleMut>);
@@ -603,8 +300,8 @@ impl<R: br::RenderPass, F: br::Framebuffer> GraphicsCommand for BeginRenderPass<
 
 pub struct NextSubpass(bool);
 impl NextSubpass {
-    const WITH_INLINE_COMMANDS: Self = Self(true);
-    const WITH_COMMAND_BUFFER_EXECUTIONS: Self = Self(false);
+    pub const WITH_INLINE_COMMANDS: Self = Self(true);
+    pub const WITH_COMMAND_BUFFER_EXECUTIONS: Self = Self(false);
 }
 impl GraphicsCommand for NextSubpass {
     fn execute(self, cb: &mut br::CmdRecord<'_, impl br::CommandBuffer + br::VkHandleMut>) {
@@ -647,26 +344,11 @@ impl GraphicsCommand for BindGraphicsDescriptorSets {
     }
 }
 
-pub struct Mesh<B: br::Buffer> {
-    pub vertex_buffers: Vec<RangedBuffer<B>>,
-    pub vertex_count: u32,
-}
-impl<B: br::Buffer> Mesh<B> {
-    pub const fn draw<'m>(&'m self, instance_count: u32) -> DrawMesh<'m, B> {
-        DrawMesh {
-            mesh: self,
-            instance_count,
-            vertex_start: 0,
-            instance_start: 0,
-        }
-    }
-}
-
 pub struct DrawMesh<'m, B: br::Buffer> {
-    mesh: &'m Mesh<B>,
-    instance_count: u32,
-    vertex_start: u32,
-    instance_start: u32,
+    pub mesh: &'m Mesh<B>,
+    pub instance_count: u32,
+    pub vertex_start: u32,
+    pub instance_start: u32,
 }
 impl<B: br::Buffer> GraphicsCommand for DrawMesh<'_, B> {
     fn execute(self, cb: &mut br::CmdRecord<'_, impl br::CommandBuffer + br::VkHandleMut>) {
